Rename files to use UpperCamelCaseCluster/Logic for localization-configuration-server
