+++ src/app/clusters/localization-configuration-server/BUILD.gn
@@ -17,8 +17,8 @@ import("//build_overrides/chip.gni")
 
 source_set("localization-configuration-server") {
   sources = [
-    "localization-configuration-cluster.cpp",
-    "localization-configuration-cluster.h",
+    "LocalizationConfigurationCluster.cpp",
+    "LocalizationConfigurationCluster.h",
   ]
 
   public_deps = [
